requirements.txt
# Core API framework (from your working version)
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning & Image Processing (from your working version)
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.0.1
opencv-python==4.8.1.78

# FAISS for similarity search (from your working version)
faiss-cpu==1.7.4

# API utilities (from your working version)
python-multipart==0.0.6
pydantic==2.4.2
requests==2.31.0

# Development tools (from your working version)
python-dotenv==1.0.0

# Standard library alternatives (from your working version)
pathlib2==2.3.7

# NEW: Additional packages for about extraction only
aiohttp==3.9.1
PyYAML==6.0.1

# NEW: Additional packages for event similarity analysis
asyncpg==0.29.0        # PostgreSQL async driver for database_service
scikit-learn==1.3.2    # Required for sklearn imports in event similarity
scipy==1.10.1          # Compatible scipy version (fixes ufunc error)
anthropic==0.7.8       # Claude API client
voyageai==0.2.1        # Voyage AI embeddings client

# Multi-LLM Provider Support (v2.1.0)
openai>=2.0.0          # OpenAI API client for ChatGPT provider (v2.x with GPT-4o, GPT-5 support)